Extract FieldNoteDetail into its own file
FieldNoteDetail is a public type and belongs alongside FieldNoteAcquisitionType
in BeastMasteryService rather than nested inside the IncubationFieldNote
service class.
